// src/app/models/MedieplanLinjer.ts
export interface MedieplanLinjer {
  Bemaerkning: string;
  BureauOrdreNr: string;
  ErWeekendGruppe:boolean;
 FarveMax: number;
  FarveMin: number;
  FarvePris: number;
  FarveRabat: number;
  FarveTillaeg: number;
  FarveTotal: number;
  ManueltAendret: boolean;
MaterialeGodtgoerelsePris: number;
MaterialeNr: string;
  MedieplanNr: number;
  MedIGrupper: string;
  MiljoeTillaeg: number;
  Mm: number;
 MmPris: number;
  MmRabat: number;
 MmTotal: number;
  MaaGiveMaterialeGodtgoerelse: boolean;
 MaaGiveMmRabat: boolean;
  NormalMmPris: number;
 PrisLaast: boolean;
  SkalGiveMaterialeGodtgoerelse: boolean;
 RabatGruppe: string;
 SendeGruppe: string;
 Total: number;
 TotalInclTillaeg: number;
TotalPris: number;
  UgeavisID: number;
 Version: number;
WebtillaegFaktureresHer: boolean;

}